Pulled admin and preferences functionality out of pages controller
