+++ app/helpers/preferences_helper.rb
@@ -1,0 +1,2 @@
+module PreferencesHelper
+end
