+++ app/views/preferences/edit.html.erb
